add dryrun/verbose options

+++ tum/management/commands/check_tum_user_state.py
@@ -10,7 +10,7 @@ def init_ldap():
     server = ldap3.Server("ldap://ads.mwn.de")
     connection = ldap3.Connection(
         server,
-        "CN=%s,OU=Users,ou=TU,ou=IAM,dc=ads,dc=mwn,dc=de" % settings.LDAP_USER,
+        f"CN={settings.LDAP_USER},OU=Users,ou=TU,ou=IAM,dc=ads,dc=mwn,dc=de",
         settings.LDAP_PASSWORD,
     )
     return connection
@@ -18,7 +18,7 @@ def init_ldap():
 
 def check_user_in_ldap(connection, uid):
     return connection.search(
-        "ou=Users,ou=TU,ou=IAM,dc=ads,dc=mwn,dc=de", "(uid=%s)" % uid
+        "ou=Users,ou=TU,ou=IAM,dc=ads,dc=mwn,dc=de", f"(uid={uid})"
     )
 
 
@@ -33,12 +33,23 @@ class Command(BaseCommand):
             default=False,
             help="Whether to delete instead of deactivate old users.",
         )
+        parser.add_argument(
+            "--dry-run",
+            action="store_true",
+            dest="dryrun",
+            default=False,
+            help="Whether to check users in LDAP directory without actually changing them.",
+        )
+        parser.add_argument(
+            "--verbose",
+            action="store_true",
+            dest="verbose",
+            default=False,
+            help="Whether to print actions to command line.",
+        )
 
     def handle(self, *args, **options):
-        if not (
-            hasattr(settings, "LDAP_USER")
-            and hasattr(settings, "LDAP_PASSWORD")
-        ):
+        if not (hasattr(settings, "LDAP_USER") and hasattr(settings, "LDAP_PASSWORD")):
             self.stdout.write(
                 "Please set LDAP_USER and LDAP_PASSWORD in configuration.py."
             )
@@ -52,23 +63,27 @@ class Command(BaseCommand):
         for sa in SocialAccount.objects.all():
             if check_user_in_ldap(connection, sa.uid):
                 if not sa.user.is_active:
-                    sa.user.is_active = True
-                    sa.user.save()
+                    if options["verbose"]:
+                        self.stdout.write(f"Activating {sa.user.username}.")
+                    if not options["dryrun"]:
+                        sa.user.is_active = True
+                        sa.user.save()
                     activate_count += 1
             elif options["delete"]:
-                sa.user.delete()
+                if options["verbose"]:
+                    self.stdout.write(f"Deleting {sa.user.username}.")
+                if not options["dryrun"]:
+                    sa.user.delete()
                 delete_count += 1
             elif sa.user.is_active:
-                sa.user.is_active = False
-                sa.user.save()
+                if options["verbose"]:
+                    self.stdout.write(f"Deactivating {sa.user.username}.")
+                if not options["dryrun"]:
+                    sa.user.is_active = False
+                    sa.user.save()
                 deactivate_count += 1
         connection.unbind()
-        self.stdout.write(
-            "Activated: %s, Deactivated: %s, Deleted: %s, Verified: %s"
-            % (
-                activate_count,
-                deactivate_count,
-                delete_count,
-                len(SocialAccount.objects.all()),
+        if not options["dryrun"]:
+            self.stdout.write(
+                f"Activated: {activate_count}, Deactivated: {deactivate_count}, Deleted: {delete_count}, Verified: {len(SocialAccount.objects.all())}"
             )
-        )
